refactor: Use hook for TabIcon logic
